Implement the topology sort in CreateCowMergeOperation
Generate a sequence of COW_COPY operations. The extents in these
operations won't be written as raw bytes in snapshots. Instead,
they will read from the source partitions.

So it's important to make sure no read after write happens on the
source partitions, similiar to the inplace update. If the topology
sort isn't possible due to cycles, we will omit some blocks to
break the cycles. And these blocks will be carried as raw bytes in
cow instead.

Bug: 162274240
Test: generate a payload, unittests pass
Change-Id: If4866704459b919d4bc09be48286b3e321b70497