update_engine: Add minimum version policy
Make the minimum version policy available in the update_engine.
This policy is set to a minimum Chrome OS version the device
must have to run.

No-update windows will not apply if the device's
version is below minimum version.

If during a Kiosk update Chrome is unreachable and we can not
determine the required Chrome OS version, we will fall back to
minimum device version and update if the device's version is older.

BUG=chromium:1117450, chromium:1084453
TEST=cros_run_unit_test --board ${BOARD} --packages update_engine

Change-Id: Ie4fc868805718c9a08a562350bfb015a70a190ac
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2388067
Tested-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Amin Hassani <ahassani@chromium.org>