update_engine: Move to C++17
The AOSP is already on C++17 and CrOS stuck on C++14 prevents them from
adding C++17 features and even if they add it, we have to revert them
when we merge. Let's enable C++17 on update_engine at least so to make
things easier.

BUG=b:163153182
TEST=CQ passes

Change-Id: I037a56fbea2b7dd8429dc3734c0b3230c87e0e55
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2382415
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Kelvin Zhang <zhangkelvin@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>