Adds connection: close to test server
The test http server does not really support connection reuse. After
each request, the connection is closed. When a url returns a redirction,
curl will try to reuse the previous connection, and fail.

This was ok before, because this kind of failure didn't count into the
total number of connection failure. But things changed in latest curl.
The max redirection is actually capped by 5, the max allowed connection
failure.

This change adds "Connection: close" header to test server. So that curl
will not try to reuse connecton.

Test: atest update_engine_unittests
Change-Id: I443c28ff3760a1cad80cfd40e354e1d0328f7561