Check for public key before checking for signatures
The current code is breaking the Chrome OS because in Chrome OS we can
install an unsigned payload into an image that doesn't have the update
payload's public key. The current code is checking for the existence of
signature first. But there is no point checking for the existence of
signature if there is not going to be a public key available to verify
it. So the order needs to be switched. This used to be the older
behavior.

Bug: 163153182
Test: cros flash
Change-Id: Ifa7026d2f288acdd4450017ce0d120272021267f