update_engine: Add enterprise rollback update policy impl
This CL adds a new rollback update policy which checks if the
update to be downloaded is a rollback by checking the value of
InstallPlan::is_rollback which is set in omaha_response_handler_action.
The intent is to consult this policy in a later CL in
ChromeOSPolicy::UpdateCanBeApplied as download time restrictions will
not be applied if current update being downloaded is a rollback.

BUG=chromium:1117450
TEST=FEATURES=test emerge-${BOARD} update_engine

Change-Id: Ib3a23a57fb60d4d0e983688f78941cdbd587718e
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2532793
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Saurabh Nijhara <snijhara@google.com>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>