update_engine: Restructure Omaha response processing
The current implementation of response processing is out of order. There
is no way to know the exact information we process from
response. Also we need to move into json based request and response and
this prepares for that. This CL moves all the response parsing logic
into a unified function. This function can be later interfaced and
implemented for other data structures.

Also there is a logic change in this CL where we set the outgoing
object's manifest version for install responses as the version that came
from the response, not the one we send into the request. Its better to
follow what the server sends rather than blindly change the version that
what we think should be. We don't really use the version for any
verification anyway so its better to be this way.

BUG=b:172219380
TEST=cros_workon_make --board reef --test update_engine

Change-Id: I7f4116d786323c045a1f86ee3d83940cc846b3e8
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2543815
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>