Calculate vbmetadigest of inactive slot
We want to improve the security of the keystore encryption key for resume
on reboot. One AI is to create the key with the vbmeta digest of the
next slot to boot into. After reboot, the decryption will fail if
ro.boot.vbmeta.digest of the new slot doesn't match the calculated value
before reboot.

Since update_engine manages the slot switch, let it calculate the digest
as well.

Bug: 177625570
Test: do a update, check that the sysprop is set.

Change-Id: I74b20fca72f6946d980b93d76990f1c8b8f246a9