Allow payload_metadata to take raw pointers with size
This way, callers of payload_metadata don't have to store their data in
std::vector. This makes parsing a `mmap()`ed payload file possible.
While we don't need the ability to parse an mmap()ed payload on device,
it is really handy for parsing payload from host-side tooling. In which
case the entire payload will be available on disk anyway, and a
streaming parse will complicate code a lot.

Test: treehugger
Change-Id: Icfbd8e849093c8f389ea802864680ab214bdd721