CheckSuperPartitionAllocatableSpace: based on device prop
Check allocatable space in super based on whether VAB is
enabled on the device, instead of basing on whether snapshot
is used for this update.

On VAB devices where snapshot is not used, e.g. secondary
update, we don't want to divide allocatable space by half.

The logic changes from:

  if (!retrofit DAP && ! update uses snapshot && !sideload)
    allocatable_space /= 2

to

  if (!retrofit DAP && ! VAB enabled)
    allocatable_space /= 2

All other changes are cosmetic.

Test: pass
Bug: 185552745
Change-Id: If182da4fdbc4b39160347205ccfa50642afc5511