update_engine: allow allocatable space to be overridden
The override is needed in cases where update_engine erroneously concludes that there isn't sufficient space for dynamic partitions, so that stock MIUI ROMs can be flashed. This patch allows it to be done via a prop (ro.dynamic.full_size).

In a typical situation where this problem occurs, you will see something like this in the recovery log:
"update_engine_sideload E 12-17 04:06:00   809   809 [ERROR:dynamic_partition_control_android.cc(830)]
The maximum size of all groups for the target slot (9126805504) has exceeded allocatable space for dynamic partitions 9125756928"

So, update_engine has used a calculation of the allocatable space that doesn't match the actual available space.
In this example, running "resetprop ro.dynamic.full_size 9126805504" via runatboot.sh or postrecoveryboot.sh will override the allocatable space and allow the ROM to flash.
If the prop is used, the override will only be triggered if the value supplied is higher than the calculated allocatable space. If the prop is not used at all, then everything will carry on as normal.

Change-Id: I54268e8166c7ca9443bcecef4fba262b20dea2d5